ci: pin chart actions to a tag (#7)
<!--
Before you open the request please review the following guidelines and
tips to help it be more easily integrated:

- Describe the scope of your change - i.e. what the change does.
- Describe any known limitations with your change.
- Please run any tests or examples that can exercise your modified code.

Thank you for contributing! I will try to test and integrate the change
as soon as I can. There is no need to bump or check in on a pull request
(it will clutter the discussion of the request).

Also don't be worried if the request is closed or not integrated.
Sometimes our priorities might not match the priorities of the pull
request. Don't fret, the open source community thrives on forks and
GitHub makes it easy to keep your changes in a forked repo.
-->

### Description of the change
<!--
Describe the scope of your change - i.e. what the change does.
Remove any sections that are not applicable.
-->

Pins the actions to a tag so Renovate will hopefully be happy to pin the
digest.

#### Removed
<!-- Any features that have been removed -->

#### Fixed
<!-- Any functionality that has been fixed -->

#### Added
<!-- Any new features that have been added -->

#### Changed
<!-- Any features that have been changed from how they were working
before -->

### Benefits
<!-- What benefits will be realized by the code change? -->

### Possible drawbacks
<!-- Describe any known limitations with your change -->

### Applicable issues
<!-- Enter any applicable Issues here (You can reference an issue using
#) -->
- fixes #

## Additional information
<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

## Checklist
<!-- [Place an '[X]' (no spaces) in all applicable fields. Please remove
unrelated fields.] -->
- [ ] Title of the PR conforms to the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) standard.
- [ ] Scope of the of the PR title contains the chart name.
- [ ] Chart version in `Chart.yaml` has been bumped according to
[Semantic Versioning](https://semver.org/).
- [ ] Chart `artifacthub.io/changes` changelog annotation has been
updated in `Chart.yaml`. See [Artifact Hub
documentation](https://artifacthub.io/docs/topics/annotations/helm/#supported-annotations)
for more info.
- [ ] Variables have been documented in the `values.yaml` file.

---------

Signed-off-by: David van der Spek <david.vanderspek@flyrlabs.com>